Refactor DAO related tests, remove transactions
Removed unnecessary transaction usage in the data access layer for http_mock, http_request_response and their respective tests. These operations do not require to be wrapped in a transaction due to their simplicity. This simplifies the code base with no loss in functionality and makes it easier to reason about the application's behavior, leading to greater maintainability.